+++ src/main/java/com/prgrms/bdbks/domain/user/controller/UserController.java
@@ -23,54 +23,21 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/users")
 @RestController
 @RequiredArgsConstructor
 public class UserController {
 	private final DefaultUserService defaultUserService;
 	private final UserMapper userMapper;
 
-	@PostMapping(value = {"/auth/signup"})
-	public ResponseEntity<String> signup(@RequestBody @Valid UserCreateRequest userCreateRequest) {
-		if (defaultUserService.findUser(userCreateRequest.getLoginId()).isPresent()) {
-			return new ResponseEntity<>("Sign Up Failed", HttpStatus.BAD_REQUEST);
-		} else {
-			this.defaultUserService.register(userCreateRequest);
-			return new ResponseEntity<>("Sign Up Completed", HttpStatus.CREATED);
-		}
-	}
-
-	@GetMapping(value = {"/users/{loginId}"})
+	@GetMapping(value = {"/{loginId}"})
 	public ResponseEntity<UserFindResponse> readUser(@PathVariable String loginId) {
 		Optional<User> user = this.defaultUserService.findUser(loginId);
 		return user.map(value -> ResponseEntity.ok(userMapper.entityToFindResponse(value)))
 			.orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
 	}
 
-	@PostMapping(value = {"/auth/login"})
-	public ResponseEntity<String> login(@RequestBody @Valid UserLoginRequest request, HttpSession session) {
-		Optional<User> user = this.defaultUserService.login(request.getLoginId(), request.getPassword());
-		if (user.isPresent()) {
-			session.setAttribute("user", user.get());
-			return new ResponseEntity<>("Login Succeeded", HttpStatus.OK);
-		}
-		return new ResponseEntity<>("Login Failed", HttpStatus.UNAUTHORIZED);
-	}
-
-	@PostMapping(value = {"/auth/logout"})
-	public ResponseEntity<Void> logout(HttpSession session) {
-		User user = (User)session.getAttribute("user");
-		if (user != null) {
-			session.removeAttribute("user");
-			session.invalidate();
-			return ResponseEntity.noContent().build();
-		} else {
-			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-		}
-
-	}
-
-	@GetMapping(value = {"/users/me"})
+	@GetMapping(value = {"/me"})
 	public ResponseEntity<UserFindResponse> getMe(HttpSession session) {
 		User user = (User)session.getAttribute("user");
 		if (user != null) {
